skills/android-automation/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "android-automation-agent"
version = "0.1.0"
description = "AI-powered Android automation agent - control any Android device using natural language. Built by Mohd Mursaleen."
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Mohd Mursaleen", email = "mohdmursaleen@gmail.com"}
]
keywords = [
    "android",
    "android automation",
    "android automation agent",
    "ai agent",
    "android ai agent",
    "adb automation",
    "adb scripting",
    "termux",
    "termux automation",
    "llm agent",
    "mobile automation",
    "android testing",
    "android bot",
    "ui automation",
    "android rpa",
    "openrouter",
    "gemini",
    "mobile testing",
    "android scripting",
    "natural language automation",
    "accessibility testing",
    "android task automation",
]
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "License :: OSI Approved :: MIT License",
    "Operating System :: Android",
    "Operating System :: POSIX :: Linux",
    "Topic :: Software Development :: Testing",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Intended Audience :: Developers",
]
dependencies = [
    "requests>=2.32.3",
    "pillow",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = ["pytest>=7.0.0", "black>=24.0.0"]

[project.urls]
Homepage = "https://github.com/Mohd-Mursaleen/android-automation-agent"
Repository = "https://github.com/Mohd-Mursaleen/android-automation-agent"
Issues = "https://github.com/Mohd-Mursaleen/android-automation-agent/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["android_agent*"]

[tool.setuptools]
py-modules = ["run", "check"]
